requirements.txt
# ML and Data
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.24.0
joblib>=1.3.0

# MLflow
mlflow>=2.9.0

# Azure ML
azure-ai-ml>=1.12.0
azure-identity>=1.15.0

# Code Quality (CI)
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
pylint>=3.0.0
mypy>=1.7.0
